web/requirements.txt
Flask==0.10.1
Jinja2==2.7
gunicorn==17.5
requests==1.2.3
gspread==0.1.0
WTForms==1.0.5
Flask-SQLAlchemy==1.0
# MySQL-python
SQLAlchemy==0.9.1
itsdangerous==0.23
markupsafe
mysql-connector-python==2.0.4
Werkzeug==0.9.4
simplejson==2.3.2
pytz==2013.8.0
